+++ handler/endpoint_test.go
@@ -8,6 +8,7 @@ import (
 	"net/http/httptest"
 	"os"
 	"strings"
+	"sync"
 	"testing"
 	"time"
 
@@ -385,7 +386,7 @@ type mockProducerResult struct {
 	rt http.RoundTripper
 }
 
-func (m *mockProducerResult) Produce(r *http.Request) chan *producer.Result {
+func (m *mockProducerResult) Produce(r *http.Request, _ *sync.Map) chan *producer.Result {
 	result := make(chan *producer.Result, 1)
 	defer close(result)
 
@@ -404,6 +405,10 @@ func (m *mockProducerResult) Produce(r *http.Request) chan *producer.Result {
 
 func (m *mockProducerResult) Len() int {
 	return 1
+}
+
+func (m *mockProducerResult) Names() []string {
+	return []string{"default"}
 }
 
 func TestEndpoint_ServeHTTP_FaultyDefaultResponse(t *testing.T) {