+++ handler/producer/request.go
@@ -30,7 +30,7 @@ type Request struct {
 // Requests represents the producer <Requests> object.
 type Requests []*Request
 
-func (r Requests) Produce(req *http.Request) chan *Result {
+func (r Requests) Produce(req *http.Request, _ *sync.Map) chan *Result {
 	var currentName string // at least pre roundtrip
 	wg := &sync.WaitGroup{}
 	ctx := req.Context()
@@ -144,6 +144,14 @@ func (r Requests) Produce(req *http.Request) chan *Result {
 
 func (r Requests) Len() int {
 	return len(r)
+}
+
+func (r Requests) Names() []string {
+	var names []string
+	for _, i := range r {
+		names = append(names, i.Name)
+	}
+	return names
 }
 
 func withRoundTripName(ctx context.Context, name string) context.Context {